Put generationlogic in a seperate project so it can be easily referenced in other project for jquery.globalize culture generation
