+++ Globalization/Resources/UmAlQuraCalendar.js
