+++ generator.sln
@@ -1,0 +1,42 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 2012
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "generator", "generator\generator.csproj", "{6EBDAB07-7C5E-4CF5-91CA-368ED36D6600}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Globalization", "Globalization\Globalization.csproj", "{B28ECA8A-87B4-43E1-A680-8648EA621DCD}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|Mixed Platforms = Debug|Mixed Platforms
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|Mixed Platforms = Release|Mixed Platforms
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{6EBDAB07-7C5E-4CF5-91CA-368ED36D6600}.Debug|Any CPU.ActiveCfg = Debug|x86
+		{6EBDAB07-7C5E-4CF5-91CA-368ED36D6600}.Debug|Mixed Platforms.ActiveCfg = Debug|x86
+		{6EBDAB07-7C5E-4CF5-91CA-368ED36D6600}.Debug|Mixed Platforms.Build.0 = Debug|x86
+		{6EBDAB07-7C5E-4CF5-91CA-368ED36D6600}.Debug|x86.ActiveCfg = Debug|x86
+		{6EBDAB07-7C5E-4CF5-91CA-368ED36D6600}.Debug|x86.Build.0 = Debug|x86
+		{6EBDAB07-7C5E-4CF5-91CA-368ED36D6600}.Release|Any CPU.ActiveCfg = Release|x86
+		{6EBDAB07-7C5E-4CF5-91CA-368ED36D6600}.Release|Mixed Platforms.ActiveCfg = Release|x86
+		{6EBDAB07-7C5E-4CF5-91CA-368ED36D6600}.Release|Mixed Platforms.Build.0 = Release|x86
+		{6EBDAB07-7C5E-4CF5-91CA-368ED36D6600}.Release|x86.ActiveCfg = Release|x86
+		{6EBDAB07-7C5E-4CF5-91CA-368ED36D6600}.Release|x86.Build.0 = Release|x86
+		{B28ECA8A-87B4-43E1-A680-8648EA621DCD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B28ECA8A-87B4-43E1-A680-8648EA621DCD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B28ECA8A-87B4-43E1-A680-8648EA621DCD}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{B28ECA8A-87B4-43E1-A680-8648EA621DCD}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{B28ECA8A-87B4-43E1-A680-8648EA621DCD}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{B28ECA8A-87B4-43E1-A680-8648EA621DCD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B28ECA8A-87B4-43E1-A680-8648EA621DCD}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B28ECA8A-87B4-43E1-A680-8648EA621DCD}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{B28ECA8A-87B4-43E1-A680-8648EA621DCD}.Release|Mixed Platforms.Build.0 = Release|Any CPU
+		{B28ECA8A-87B4-43E1-A680-8648EA621DCD}.Release|x86.ActiveCfg = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
